requirements.txt
flet==0.28.3
cryptography
py-diffie-hellman
pyjwt
yarl